+++ v2/components/UserBalances/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@snx-v2/UserBalances",
+  "private": true,
+  "main": "index.tsx",
+  "version": "0.0.0",
+  "dependencies": {
+    "@chakra-ui/react": "^2.2.8",
+    "@snx-v2/icons": "workspace:*",
+    "@synthetixio/wei": "workspace:*"
+  }
+}
